fix: keep defaults lifecycle test out of CEC
